requirements.txt
imutils==0.5.4
numpy==1.25.1
opencv-python==4.8.0.74
Pillow==10.0.0
